service/requirements.txt
python-nmap
Pyside6
psutil